requirements.txt
Django==5.0.7
psycopg[binary]==3.2.1
pytz==2024.1
sqlparse==0.5.1

# REST Framework
djangorestframework==3.15.2

# JWT Auth
djangorestframework-simplejwt==5.3.1

# Filtering and Search
django-filter==24.2

# API Docs
drf-yasg==1.21.7

# CORS
django-cors-headers==4.4.0

# Database drivers
mysqlclient==2.2.4  # For MySQL on PythonAnywhere

# HTTP requests for geolocation services
requests==2.32.3

# Redis (caching / future rate limiting)
redis==5.0.8
django-redis==5.4.0

# Background task processing (optional for self-hosting)
celery==5.3.4
django-celery-beat==2.6.0
django-celery-results==2.6.0

# Excel export support for analytics
openpyxl==3.1.5

# IP address utilities for compliance
ipaddress==1.0.23

# JSON field improvements for older Django versions
django-jsonfield==1.4.1

# Timezone utilities
pytz==2024.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0  # For serving static files

# Dev
black==24.8.0
isort==5.13.2

# Env variables
python-dotenv==1.0.1

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0  # For test data generation
pytest-mock==3.14.0
pytest-cov==5.0.0